src/test/java/com/cloud/taco/project/controllers/OrderControllerTest.java
package com.cloud.taco.project.controllers;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.ui.Model;

import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@ExtendWith(MockitoExtension.class)
class OrderControllerTest {

    @Mock
    Model model;

    @InjectMocks
    OrderController controller;


    @Test
    void shouldReturnIndex() throws Exception {

        MockMvc mockMvc = MockMvcBuilders.standaloneSetup(controller).build();

        mockMvc.perform(get("/order"))
                .andExpect(status().isOk())
                .andExpect(view().name("order/index"));

    }

    @Test
    void modelShouldBeExecutedOnceForEachAttribute() {

        String viewName = controller.getForm(model);
        verify(model, times(1)).addAttribute(eq("order"), any());

    }
}